requirements.txt
torch>=2.0.0
torchvision
transformers>=4.30.0
pandas
Pillow
matplotlib
numpy
scikit-learn
